for log_processor testing
